Updates `network/gauge_network.py`
More specifically, removes `network/gauge_network.py` and renames
`network/gauge_network_new.py` to `network/gauge_network.py`
